pyproject.toml
[tool.poetry]
name = "codedancer"
version = "0.1.0"
description = ""
authors = ["svtter <svtter@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gitpython = "^3.1.40"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
